Updated translations for zh-Hans, zh-Hant and zh-HK
"精算" is a hard one to translate. While "精算" is kanji and you can sort of infer the vague meaning in Chinese, and a lot of people who travel to Japan frequently knows what it means, the true meaning of "精算" in Chinese still doesn't align with "adjustment". The everyday language equivelant of "精算" in Chinese is "補票", which literally means "to add [something missing] to your ticket", i.e. to make up the difference in price for your ticket. This is the easiest to understand but it doesn't contain the "adjustment" part of the the original Japanese word. So for the sake of respecting the original meaning, while also maintaining some level of clarity, I've opted to go for the literal translation, "車費調整" for "Fare Adjustment", and "票價調整" for "Ticket Adjustment".